Reglage conflit

+++ Code/spring.boot.planning/src/main/java/com/hitema/spring/boot/planning/entities/Session.java
@@ -12,7 +12,7 @@ public class Session {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "session_id")
-    private Integer id;
+    private Long id;
     @Column(name = "nom")
     private String nom ;
     @Column(name = "unites")
@@ -20,6 +20,7 @@ public class Session {
 
     @Column(name="date_creation")
     private LocalDateTime creationDate;
+
 
     private Integer ResultatVote;
     @ManyToMany(mappedBy = "sessions")
@@ -37,11 +38,11 @@ public class Session {
                 '}';
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
